% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_beta_diversity.R
\name{calculate_beta_diversity}
\alias{calculate_beta_diversity}
\title{Calculate Beta Diversity}
\usage{
calculate_beta_diversity(
  abund_table,
  metric = "bray-curtis",
  output_dir = NULL,
  file_name = "beta_diversity_matrix.csv"
)
}
\arguments{
\item{abund_table}{A microbiome abundance table (matrix or data frame)}

\item{metric}{Beta diversity metric to use: "bray-curtis" or "jaccard"}

\item{output_dir}{Optional; a character string specifying the file path to save the output. If provided, the output will be saved to this path. Default is NULL.}

\item{file_name}{Optional; a character string specifying the name of the file to save the plot as. Default is "beta_diversity_matrix.csv".}
}
\value{
A list containing:
\item{result_mat}{A distance object}
\item{result_df}{A dataframe representing a NxN matrix with the distances, which is the one saved in output}
}
\description{
This function calculates beta diversity for a given microbiome table.
}
