% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_to_factors.R
\name{convert_columns_to_factors}
\alias{convert_columns_to_factors}
\title{Convert Specified Columns to Factors}
\usage{
convert_columns_to_factors(df, columns)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{columns}{A character vector of column names to be checked and possibly converted to factors.}
}
\value{
The modified data frame with specified columns converted to factors.
}
\description{
This function takes a data frame and a vector of column names, and converts
the specified columns to factors if they are not already factors or numeric.
}
