% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{cowplot_title}
\alias{cowplot_title}
\title{Add title to a cowplot object}
\usage{
cowplot_title(p, title, rel_heights = c(0.1, 1))
}
\arguments{
\item{p}{the cowplot object}

\item{title}{title}

\item{rel_heights}{relative heights between the title and the cowplot object}
}
\value{
the plot with title added in
}
\description{
Add title to a cowplot object
}
