% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Generate a color scheme similar to ggplot's default, given either the number of categories
or the categories themselves}
\usage{
gg_color_hue(values = NULL, n = NULL)
}
\arguments{
\item{values}{separate categories, or}

\item{n}{the number of categories}
}
\value{
length-n color scheme
}
\description{
Generate a color scheme similar to ggplot's default, given either the number of categories
or the categories themselves
}
