% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqs.R
\name{kOverA2}
\alias{kOverA2}
\title{Filtering utility to indicate if a feature is > A present in at least k samples}
\usage{
kOverA2(k = 1, A = 0)
}
\arguments{
\item{k}{how many samples should the feature be present in?}

\item{A}{abundance threshold for the feature to be considered as present}
}
\description{
Filtering utility to indicate if a feature is > A present in at least k samples
}
