% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maas_consistency.R
\name{maas_consistency}
\alias{maas_consistency}
\title{Compare and Process MaAsLin2 and MaAsLin3 Results}
\usage{
maas_consistency(
  df_maas2_,
  df_maas3_,
  top_features_,
  key_cols_,
  coef_cols_,
  stderr_cols_,
  qval_cols_
)
}
\arguments{
\item{df_maas2_}{Data frame containing MaAsLin2 results.}

\item{df_maas3_}{Data frame containing MaAsLin3 results.}

\item{top_features_}{Integer. Number of top features to select based on coefficient magnitude.}

\item{key_cols_}{Character vector. Names of columns to use as keys for merging.}

\item{coef_cols_}{Character vector. Names of coefficient columns.}

\item{stderr_cols_}{Character vector. Names of standard error columns.}

\item{qval_cols_}{Character vector. Names of q-value columns.}
}
\value{
A list containing three data frames:
\item{df_maas_}{Merged and filtered results from both analyses}
\item{df_maas_top_}{Top features based on coefficient magnitude}
\item{df_maas_plot_}{Data prepared for plotting}
}
\description{
This function merges and processes the results from MaAsLin2 and MaAsLin3 analyses,
identifying consistent features and preparing data for visualization.
}
\examples{
df_maas2 <- data.frame(feature = c("A", "B", "C"), coef.x = c(1, -2, 3),
                       stderr.x = c(0.1, 0.2, 0.3), qval.x = c(0.01, 0.02, 0.03))
df_maas3 <- data.frame(feature = c("B", "C", "D"), coef.y = c(-2.1, 3.1, 4),
                       stderr.y = c(0.15, 0.25, 0.35), qval.y = c(0.015, 0.025, 0.035))
result <- maas_consistency(df_maas2, df_maas3, 2, "feature",
                           c("coef.x", "coef.y"), c("stderr.x", "stderr.y"),
                           c("qval.x", "qval.y"))

}
