% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqs.R
\name{otu_table2}
\alias{otu_table2}
\title{Replaces phyloseq::otu_table and returns a matrix instead}
\usage{
otu_table2(physeq)
}
\arguments{
\item{physeq}{}
}
\value{
matrix of feature table
}
\description{
Replaces phyloseq::otu_table and returns a matrix instead
}
