% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqs.R
\name{phyloseq_to_tb}
\alias{phyloseq_to_tb}
\title{Create long format tibble object given phyloseq, aggregating otu table (must have),
metadata (optional), and tax table (optional) together.}
\usage{
phyloseq_to_tb(physeq)
}
\arguments{
\item{physeq}{}
}
\value{
a long tibble, each row corresponding to one feature and one sample
}
\description{
Create long format tibble object given phyloseq, aggregating otu table (must have),
metadata (optional), and tax table (optional) together.
}
