% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_diversity.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot Alpha Diversity with Optional Faceting}
\usage{
plot_alpha_diversity(
  alpha_table,
  metadata_table,
  plot_type = "boxplot",
  plot_variables,
  title = NULL,
  output_dir = NULL,
  file_name = "alpha_diversity_plot.png"
)
}
\arguments{
\item{alpha_table}{A data frame (or matrix) where the rownames are samples and the first column contains the alpha diversity measure to plot.}

\item{metadata_table}{A data frame where the rownames are samples and the columns contain metadata information.}

\item{plot_type}{A character string specifying the type of plot to create: "boxplot" (default) or "violinplot".}

\item{plot_variables}{A character string or vector specifying the name(s) of the column(s) to use for separating the samples on the x-axis of the resulting plot. If more than one column is provided, faceting will be applied.}

\item{title}{Optional; a character string specifying the title of the plot. Default is NULL.}

\item{output_dir}{Optional; a character string specifying the directory where the plot should be saved. If provided, the directory will be created if it does not exist.}

\item{file_name}{Optional; a character string specifying the name of the file to save the plot as. Default is "alpha_diversity_plot.png".}
}
\value{
A ggplot2 plot object.
}
\description{
This function creates a boxplot or violin plot from a data frame of alpha diversity measures.
If multiple metadata columns are provided, it will automatically apply faceting.
}
