% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_diversity.R
\name{plot_beta_diversity}
\alias{plot_beta_diversity}
\title{Perform NMDS on Beta Diversity Matrix and Plot Results}
\usage{
plot_beta_diversity(
  beta_mat,
  metadata_table,
  plot_variables = NULL,
  maxit = 50,
  output_dir = NULL,
  file_name = "beta_diversity_plot.png"
)
}
\arguments{
\item{beta_mat}{A distance matrix representing beta diversity, either as a matrix or a dist object.}

\item{metadata_table}{A data frame containing metadata information, where the rownames match the sample names in beta_mat.}

\item{plot_variables}{A character string specifying the metadata columns to show in the plot. Default is NULL.
Can be either one variable or a list with two variables.}

\item{maxit}{An integer specifying the maximum number of iterations for the NMDS algorithm (default is 50).}

\item{output_dir}{Optional; a character string specifying the directory where the plot should be saved. If provided, the directory will be created if it does not exist.}

\item{file_name}{Optional; a character string specifying the name of the file to save the plot as. Default is "beta_diversity_plot.png".}
}
\value{
A ggplot2 plot object.
}
\description{
This function calculates a 2D NMDS (Non-metric Multidimensional Scaling) from a beta diversity distance matrix
and generates a scatter plot of the resulting points, optionally coloring the points based on a metadata variable.
}
