% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_maaslin_results.R
\name{process_maaslin_results}
\alias{process_maaslin_results}
\title{Process MaAsLin Results}
\usage{
process_maaslin_results(df_maas_, pthresh_maas, qthresh_maas, required_cols)
}
\arguments{
\item{df_maas_}{A data frame containing the MaAsLin results.}

\item{pthresh_maas}{Numeric. The p-value threshold for filtering results. (0 < p-value <= q-value)}

\item{qthresh_maas}{Numeric. The q-value threshold for filtering results. (p-value <= q-value <= 1)}

\item{required_cols}{Character vector. Names of columns that must be present and non-NA.}
}
\value{
A data frame containing the filtered and processed MaAsLin results.
}
\description{
This function filters and processes the results from a MaAsLin analysis based on
specified p-value and q-value thresholds. It also ensures that all required columns
are present and contain non-NA values.
}
