% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqs.R
\name{prune_taxaSamples}
\alias{prune_taxaSamples}
\title{Prune a phyloseq object to exclude empty taxa/samples, which
are prone to happen when a phyloseq object is subsetted.}
\usage{
prune_taxaSamples(physeq, flist_taxa = kOverA2(k = 1, A = 0),
  flist_samples = function(x) sum(x > 0) > 0, max.iter = 3)
}
\arguments{
\item{flist_taxa}{function for filtering taxa, default to non-empty pruning}

\item{flist_samples}{function for filtering sample, default to non-empty pruning}

\item{max.iter}{maximum number of iterations}
}
\value{
the pruned phyloseq object
}
\description{
Prune a phyloseq object to exclude empty taxa/samples, which
are prone to happen when a phyloseq object is subsetted.
}
