% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{rotate_xaxis}
\alias{rotate_xaxis}
\title{Rotate x axis of a ggplot}
\usage{
rotate_xaxis(angle, hjust = 1, vjust = 1)
}
\arguments{
\item{angle}{angle to rotate}

\item{hjust}{hjust}

\item{vjust}{vjust}
}
\value{
plot with x axis rotated as specified
}
\description{
Rotate x axis of a ggplot
}
