% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_maaslin2.R
\name{run_maaslin2}
\alias{run_maaslin2}
\title{Run MaAsLin2 Analysis}
\usage{
run_maaslin2(
  abund_table,
  metadata_table,
  fixed_effects,
  random_effects = NULL,
  other_params = list(),
  output_dir,
  rowtype = "feature"
)
}
\arguments{
\item{abund_table}{A matrix or data frame of abundance data}

\item{metadata_table}{A data frame of metadata}

\item{fixed_effects}{A character vector of column names in metadata to use as fixed effects}

\item{random_effects}{A character vector of column names in metadata to use as random effects}

\item{other_params}{A list of other parameters to pass to MaAsLin2}

\item{output_dir}{The directory to save the MaAsLin2 output}

\item{rowtype}{The type of data in the rows of abund_table ("feature" or "sample")}
}
\value{
The results of the MaAsLin2 analysis
}
\description{
This function runs a MaAsLin2 analysis on the provided abundance table and metadata.
}
