% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqs.R
\name{sample_data2}
\alias{sample_data2}
\title{Replaces phyloseq::sample_data and returns a data frame instead}
\usage{
sample_data2(physeq)
}
\arguments{
\item{physeq}{}
}
\value{
data frame of metadata
}
\description{
Replaces phyloseq::sample_data and returns a data frame instead
}
