% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{set_ggplot}
\alias{set_ggplot}
\title{My preferred ggplot theme}
\usage{
set_ggplot(title.size = 16, axis.title.size = 15,
  axis.text.size = 14, legend.text.size = 14, strip.text.size = 14)
}
\arguments{
\item{title.size}{font size for title}

\item{axis.title.size}{font size for axis title}

\item{axis.text.size}{font size for axis text}

\item{legend.text.size}{font size for legend}

\item{strip.text.size}{font size for grid panels}
}
\value{
does not return anything
}
\description{
My preferred ggplot theme
}
