% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Source all R scripts under a directory}
\usage{
sourceDir(path, pattern = "[.][Rr]$", trace = TRUE,
  recursive = FALSE, ...)
}
\arguments{
\item{trace}{should list of files be outputted?}

\item{...}{}
}
\value{
nothing
}
\description{
Source all R scripts under a directory
}
