% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_distance.R
\name{subset_distance}
\alias{subset_distance}
\title{Subsetting a distance object by indicators}
\usage{
subset_distance(dist, ind)
}
\arguments{
\item{dist}{an n x n distance object}

\item{ind}{length-n subsetting indicator. Must be either character (sample names) or
logical (TRUE/FALSE indicator).}
}
\value{
the subsetted distance object
}
\description{
Subsetting a distance object by indicators
}
