% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqs.R
\name{tax_table2}
\alias{tax_table2}
\title{Replaces phyloseq::tax_table and returns a matrix instead}
\usage{
tax_table2(physeq)
}
\arguments{
\item{physeq}{}
}
\value{
matrix of taxonamy table
}
\description{
Replaces phyloseq::tax_table and returns a matrix instead
}
