% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_abund_table.R
\name{validate_abund_table}
\alias{validate_abund_table}
\title{Validate and Preprocess Microbiome Relative Abundance Tables as Input}
\usage{
validate_abund_table(abund_table, rowtype)
}
\arguments{
\item{abund_table}{A microbiome abundance table (matrix or data frame). Rows are species, and columns are samples.}

\item{rowtype}{A character string specifying the type of data in rows (e.g., "species", "OTUs" or "samples").}
}
\value{
The processed abund_table with transposed data if needed, samples with zero profiles removed, and normalized values.
}
\description{
This function performs validation and correction on a relative abundance table, following common preprocessing steps on inputs of a computational metagenomic pipeline.
It includes steps to handle zero-abundance rows and columns, check and correct table orientation, and normalize column values.
}
\details{
This function validates and preprocesses the inputs of a computational metagenomic pipeline.
}
